move functions around to make code more dry and consistent
